Tested current DB and preparing for merge
