--- src/main/java/TeamSun/CS4800Project/dao/TestDao.java
@@ -1,6 +1,0 @@
-package TeamSun.CS4800Project.dao;
-
-public interface TestDao {
-	//TODO implement
-
-}
