--- src/main/java/TeamSun/CS4800Project/dao/TestDaoDB.java
@@ -1,9 +1,0 @@
-package TeamSun.CS4800Project.dao;
-
-import org.springframework.stereotype.Repository;
-
-@Repository("TestDB")
-public class TestDaoDB implements TestDao {
-	//TODO implement
-
-}
