--- src/main/java/TeamSun/CS4800Project/repositories/TestRepo.java
@@ -1,14 +1,0 @@
-package TeamSun.CS4800Project.repositories;
-
-import java.util.Optional;
-
-import org.bson.types.ObjectId;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import TeamSun.CS4800Project.model.TestModel;
-
-public interface TestRepo extends MongoRepository<TestModel, ObjectId> {
-
-	public Optional<TestModel> findById(ObjectId id);
-	
-}
