+++ src/main/java/TeamSun/CS4800Project/service/PersonService.java
@@ -36,5 +36,10 @@ public class PersonService {
 	public int size() {
 		return DB.size();
 	}
+	
+	@Override
+	public String toString() {
+		return DB.getAll().toString();
+	}
 
 }
