--- src/main/java/TeamSun/CS4800Project/service/TestService.java
@@ -1,7 +1,0 @@
-package TeamSun.CS4800Project.service;
-
-public class TestService {
-	
-	//TODO implement
-
-}
